+++ components/LoginView.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import {
   StyleSheet,
   Text,
@@ -10,103 +10,52 @@ import {
   ImageBackground,
   Alert
 } from 'react-native';
-import token from '../token';
-import { AuthSession } from 'expo';
-const client_id = 'f7410f08c2064e4c9517603f56ed4089';
 
-export default class LoginView extends React.Component {
+LoginView = (props) => {
+  return (
+    <View style={styles.container}>
+    <ImageBackground source={require('../img/vinylSplash.jpg')} style={styles.imgBackground}>
+      <Text style={styles.headerText}>Welcome</Text>
+      <View style={styles.inputContainer}>
+        <Image style={styles.inputIcon} source={{uri: 'https://png.icons8.com/message/ultraviolet/50/3498db'}}/>
+        <TextInput style={styles.inputs}
+            placeholder="Email"
+            keyboardType="email-address"
+            underlineColorAndroid='transparent'
+            onChangeText={(email) => this.setState({email})}/>
+      </View>
 
-  constructor(props) {
-    super(props);
-    state = {
-      email   : '',
-      password: '',
-    }
-  }
+      <View style={styles.inputContainer}>
+        <Image style={styles.inputIcon} source={{uri: 'https://png.icons8.com/key-2/ultraviolet/50/3498db'}}/>
+        <TextInput style={styles.inputs}
+            placeholder="Password"
+            secureTextEntry={true}
+            underlineColorAndroid='transparent'
+            onChangeText={(password) => this.setState({password})}/>
+      </View>
 
-  onClickListener = (viewId) => {
-    Alert.alert("Alert", "Button pressed " + viewId);
-  }
+      <TouchableHighlight style={[styles.buttonContainer, styles.loginButton]} onPress={() => props.navigation.navigate('ProfileScreen')}>
+        <Text style={styles.loginText}>Login</Text>
+      </TouchableHighlight>
 
-  generateRandomString = (length) => {
-    let text = '';
-    let possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+      <TouchableHighlight style={[styles.buttonContainer, styles.spotifyButton]} onPress={() => props.spotifyLogin()}>
+        <Text style={styles.spotifyText}>Spotify</Text>
+      </TouchableHighlight>
 
-    for (let i = 0; i < length; i++) {
-      text += possible.charAt(Math.floor(Math.random() * possible.length));
-    }
-    return text;
-  };
+      <TouchableHighlight style={[styles.buttonContainer, styles.googleButton]} onPress={() => props.navigation.navigate('AltLogin')}>
+        <Text style={styles.googleText}>Google</Text>
+      </TouchableHighlight>
 
-  spotifyLogin = async() => {
-    let redirectUrl = AuthSession.getRedirectUrl();
-    let scope = 'user-library-read user-top-read';
-    let state = this.generateRandomString(16);
+      <TouchableHighlight style={styles.buttonContainer} onPress={() => this.onClickListener('restore_password')}>
+          <Text style={styles.sundryText}>Forgot your password?</Text>
+      </TouchableHighlight>
 
-    let result = await AuthSession.startAsync({
-      authUrl:
-        'https://accounts.spotify.com/authorize?' +
-        '&response_type=code' +
-        '&client_id=' + client_id +
-        (scope ? '&scope' + encodeURIComponent(scope) : '') +
-        '&redirect_uri=' + encodeURIComponent(redirectUrl) +
-        '&state=' + state
-    });
-
-    if(result.type !== 'success'){
-      Alert.alert('Spotify login unsuccessful');
-      return;
-    }
-    const newToken = await token(result.params.code, redirectUrl);
-    this.props.navigation.navigate('SpotifyInitial');
-  }
-
-  render() {
-      return (
-        <View style={styles.container}>
-        <ImageBackground source={require('../img/vinylSplash.jpg')} style={styles.imgBackground}>
-          <Text style={styles.headerText}>Welcome</Text>
-          <View style={styles.inputContainer}>
-            <Image style={styles.inputIcon} source={{uri: 'https://png.icons8.com/message/ultraviolet/50/3498db'}}/>
-            <TextInput style={styles.inputs}
-                placeholder="Email"
-                keyboardType="email-address"
-                underlineColorAndroid='transparent'
-                onChangeText={(email) => this.setState({email})}/>
-          </View>
-
-          <View style={styles.inputContainer}>
-            <Image style={styles.inputIcon} source={{uri: 'https://png.icons8.com/key-2/ultraviolet/50/3498db'}}/>
-            <TextInput style={styles.inputs}
-                placeholder="Password"
-                secureTextEntry={true}
-                underlineColorAndroid='transparent'
-                onChangeText={(password) => this.setState({password})}/>
-          </View>
-
-          <TouchableHighlight style={[styles.buttonContainer, styles.loginButton]} onPress={() => this.props.navigation.navigate('ProfileScreen')}>
-            <Text style={styles.loginText}>Login</Text>
-          </TouchableHighlight>
-
-          <TouchableHighlight style={[styles.buttonContainer, styles.spotifyButton]} onPress={() => this.spotifyLogin()}>
-            <Text style={styles.spotifyText}>Spotify</Text>
-          </TouchableHighlight>
-
-          <TouchableHighlight style={[styles.buttonContainer, styles.googleButton]} onPress={() => this.props.navigation.navigate('AltLogin')}>
-            <Text style={styles.googleText}>Google</Text>
-          </TouchableHighlight>
-
-          <TouchableHighlight style={styles.buttonContainer} onPress={() => this.onClickListener('restore_password')}>
-              <Text style={styles.sundryText}>Forgot your password?</Text>
-          </TouchableHighlight>
-
-          <TouchableHighlight style={styles.buttonContainer} onPress={() => this.onClickListener('register')}>
-              <Text style={styles.sundryText}>Register</Text>
-          </TouchableHighlight>
-          </ImageBackground>
-        </View>
-      );
-  }
+      <TouchableHighlight style={styles.buttonContainer} onPress={() => this.onClickListener('register')}>
+          <Text style={styles.sundryText}>Register</Text>
+      </TouchableHighlight>
+      </ImageBackground>
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
